added inheritance task example
This task Object-Oriented Programming task demonstrates inheritance
from a base/parent class to a derived/child class and overriding a method - hence the title 'method_override'

For this example, we gather information about an individual, create a class object with their details and
determine whether they are eligible to write a learner's test. We then output a fitting message based on those details
while accounting for error inputs.